+++ Utils.h
@@ -18,5 +18,6 @@ class Utils {
 public:
     static void string_to_bit_stream(std::vector<bool> &vector, std::string message);
     static void bit_string_to_bit_stream(std::vector<bool> &vector, std::string bitstring);
+    static std::string zero_pad_int(int num, int totalLength);
 };
 #endif //SAMEENCODER_UTILS_H
